requirements.txt
Babel==2.14.0
docxcompose==1.4.0
docxtpl==0.16.7
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
python-docx==1.1.0
six==1.16.0
typing_extensions==4.9.0